Show Account Usage

Show account usage

accountUsage(
  account = NULL,
  server = NULL,
  usageType = "hours",
  from = NULL,
  until = NULL,
  interval = NULL
)

Arguments

account

Account name. If a single account is registered on the system then this parameter can be omitted.

server

Server name. Required only if you use the same account name on multiple servers.

usageType

Use metric to retreive (for example: "hours")

from

Date range starting timestamp (Unix timestamp or relative time delta such as "2d" or "3w").

until

Date range ending timestamp (Unix timestamp or relative time delta such as "2d" or "3w").

interval

Summarization interval. Data points at intervals less then this will be grouped. (Number of seconds or relative time delta e.g. "1h").

Note

This function only works for ShinyApps servers.