accounts {rsconnect}	R Documentation

Account Management Functions

Description

Functions to enumerate and remove accounts on the local system. Prior to deploying applications you need to register your account on the local system.

Usage

accounts(server = NULL)

accountInfo(name, server = NULL)

removeAccount(name, server = NULL)

Arguments

server
Name of the server on which the account is registered (optional; see servers())

name
Name of account

Details

You register an account using the setAccountInfo() function (for ShinyApps) or connectUser() function (for other servers). You can subsequently remove the account using the removeAccount function.

The accounts and accountInfo functions are provided for viewing previously registered accounts.

Value

accounts returns a data frame with the names of all accounts registered on the system and the servers on which they reside. accountInfo returns a list with account details.

See Also

Other Account functions: connectApiUser(), connectUser(), setAccountInfo()

[Package rsconnect version 0.8.26 Index]