applicationConfigDir {rsconnect}	R Documentation

Application Configuration Directory

Description

Returns the root path used to store per user configuration data. Does not check old locations or create the path; use rsconnectConfigDir for most cases.

Usage

applicationConfigDir()

Value

A string containing the path of the configuration folder.

[Package rsconnect version 0.8.26 Index]