List Deployed Applications

List all applications currently deployed for a given account.

applications(account = NULL, server = NULL)

Arguments

account
Account name. If a single account is registered on the system then this parameter can be omitted.

server
Server name. Required only if you use the same account name on multiple servers.

Value

Returns a data frame with the following columns:

id	Application unique id

name	Name of application

url	URL where application can be accessed

status	Current status of application. Valid values are pending, deploying, running, terminating, and terminated

size	Instance size (small, medium, large, etc.) (on ShinyApps.io)

instances	Number of instances (on ShinyApps.io)

config_url	URL where application can be configured

Note

To register an account you call the setAccountInfo() function.

See also

Examples

if (FALSE) {

# list all applications for the default account
applications()

# list all applications for a specific account
applications("myaccount")

# view the list of applications in the data viewer
View(applications())
}