connectUser {rsconnect}	R Documentation

Connect User Account

Description

Connect a user account to the package so that it can be used to deploy and manage applications on behalf of the account.

Usage

connectUser(account = NULL, server = NULL, quiet = FALSE)

Arguments

account
A name for the account to connect. Optional.

server
The server to connect to. Optional if there is only one server registered.

quiet
Whether or not to show messages and prompts while connecting the account.

Details

When this function is invoked, a web browser will be opened to a page on the target server where you will be prompted to enter your credentials. Upon successful authentication, your local installation of rsconnect and your server account will be paired, and you'll be able to deploy and manage applications using the package without further prompts for credentials.

See Also

Other Account functions: accounts(), connectApiUser(), setAccountInfo()

[Package rsconnect version 0.8.26 Index]