deployDoc {rsconnect}	R Documentation

Deploy a single document

Description

Deploys a single R Markdown, Quarto document, or other file (e.g. .html or .pdf).

When deploying an .Rmd, .Qmd, or .html, deployDoc() will attempt to automatically discover dependencies using rmarkdown::find_external_resources(), and include an .Rprofile if present. If you find that the document is missing dependencies, either specify the dependencies explicitly in the document (see rmarkdown::find_external_resources() for details), or call deployApp() directly and specify your own file list in appFiles.

Usage

deployDoc(doc, ..., logLevel = c("normal", "quiet", "verbose"))

Arguments

doc
Path to the document to deploy.

...
Additional arguments to deployApp(). Do not supply appDir, appFiles, or appPrimaryDoc; these three parameters are automatically generated by deployDoc from the document.

logLevel
One of "quiet", "normal" or "verbose"; indicates how much logging to the console is to be performed. At "quiet" reports no information; at "verbose", a full diagnostic log is captured.

See Also

Other Deployment functions: applications(), deployAPI(), deployApp(), deploySite(), deployTFModel()

Examples

## Not run: 
deployDoc("my-report.Rmd")
deployDoc("static-file.html")

## End(Not run)

[Package rsconnect version 1.2.1 Index]