lint {rsconnect}	R Documentation

Lint a Project

Description

Takes the set of active linters (see addLinter()), and applies them to all files within a project.

Usage

lint(project, files = NULL, appPrimaryDoc = NULL)

Arguments

project
Path to a project directory.

files
Specific files to lint. Can be NULL, in which case all the files in the directory will be linted.

appPrimaryDoc
The primary file in the project directory. Can be NULL, in which case it's inferred (if possible) from the directory contents.

[Package rsconnect version 0.8.26 Index]