listBundleFiles {rsconnect}	R Documentation

List Files to be Bundled

Description

listBundleFiles() has been superseded in favour of listDeploymentFiles().

Given a directory containing an application, returns the names of the files that by default will be bundled in the application. It works similarly to a recursive directory listing from list.files() but enforces bundle sizes as described in listDeploymentFiles()

Usage

listBundleFiles(appDir)

Arguments

appDir
Directory containing the application.

Value

Returns a list containing the following elements:

totalFiles: Total number of files.

totalSize: Total size of the files (in bytes).

contents: Paths to bundle, relative to appDir.

[Package rsconnect version 1.2.1 Index]