oldApplicationConfigDir {rsconnect}	R Documentation

Old Application Config Directory

Description

Returns the old application configuration directory used by rsconnect 0.8.24 and prior. These versions wrote configuration data to XDG compliant locations, but CRAN policy has since further restricted the disk locations that are permitted. See:

Usage

oldApplicationConfigDir(appName)

Arguments

appName
The application's name (connect or rsconnect)

Details

https://cran.r-project.org/web/packages/policies.html

Value

The old application configuration directory.

[Package rsconnect version 0.8.26 Index]