rsconnectConfigDir {rsconnect}	R Documentation

rsconnect Configuration Directory

Description

Forms the path to a location on disk where user-level configuration data for the package is stored.

Usage

rsconnectConfigDir(subDir = NULL)

Arguments

subDir
An optional subdirectory to be included as the last element of the path.

Value

The path to the configuration directory.

[Package rsconnect version 0.8.26 Index]