Server Management Functions

Functions to manage the list of known servers to which rsconnect can deploy and manage applications.

servers(local = FALSE)

discoverServers(quiet = FALSE)

addConnectServer(url, name = NULL, certificate = NULL, quiet = FALSE)

addServer(url, name = NULL, certificate = NULL, quiet = FALSE)

removeServer(name)

serverInfo(name)

addServerCertificate(name, certificate, quiet = FALSE)

Arguments

local

Return only local servers (i.e. not shinyapps.io)

quiet

Suppress output and prompts where possible.

url

Server's URL. Should look like http://servername/ or http://servername:port/.

name

Optional nickname for the server. If none is given, the nickname is inferred from the server's hostname.

certificate

Optional; a path a certificate file to be used when making SSL connections to the server. The file's contents are copied and stored by the rsconnect package. Can also be a character vector containing the certificate's contents.

Value

servers returns a data frame with registered server names and URLs. serverInfo returns a list with details for a particular server.

Details

Register a server with addServer or discoverServers (the latter is useful only if your administrator has configured server autodiscovery). Once a server is registered, you can connect to an account on the server using connectUser().

The servers and serverInfo functions are provided for viewing previously registered servers.

There is always at least one server registered (the shinyapps.io server)

Examples

if (FALSE) {

# register a local server
addServer("http://myrsconnect/", "myserver")

# list servers
servers(local = TRUE)

# connect to an account on the server
connectUser(server = "myserver")
}