showProperties {rsconnect}	R Documentation

Show Application property

Description

Show properties of an application deployed to ShinyApps.

Usage

showProperties(
  appPath = getwd(),
  appName = NULL,
  account = NULL,
  server = NULL
)

Arguments

appPath
Directory or file that was deployed. Defaults to current working directory.

appName
Name of application

account, server
Uniquely identify a remote server with either your user account, the server name, or both. If neither are supplied, and there are multiple options, you'll be prompted to pick one.

Use accounts() to see the full list of available options.

Note

This function works only for ShinyApps servers.

[Package rsconnect version 1.2.1 Index]