showUsage {rsconnect}	R Documentation

Show Application Usage

Description

Show application usage of a currently deployed application

Usage

showUsage(
  appDir = getwd(),
  appName = NULL,
  account = NULL,
  server = NULL,
  usageType = "hours",
  from = NULL,
  until = NULL,
  interval = NULL
)

Arguments

appDir
Directory containing application. Defaults to current working directory.

appName
Name of application

account
Account name. If a single account is registered on the system then this parameter can be omitted.

server
Server name. Required only if you use the same account name on multiple servers.

usageType
Use metric to retreive (for example: "hours")

from
Date range starting timestamp (Unix timestamp or relative time delta such as "2d" or "3w").

until
Date range ending timestamp (Unix timestamp or relative time delta such as "2d" or "3w").

interval
Summarization interval. Data points at intervals less then this will be grouped. (Relative time delta e.g. "120s" or "1h" or "30d").

Note

This function only works for ShinyApps servers.

[Package rsconnect version 0.8.26 Index]