syncAppMetadata {rsconnect}	R Documentation

Sync Application Metadata

Description

Update the metadata for requested application across all deployments

Usage

syncAppMetadata(appPath)

Arguments

appPath
The path to the directory or file that was deployed.

Note

This function does not update metadata for Shiny and rpubs apps

[Package rsconnect version 0.8.26 Index]