taskLog {rsconnect}	R Documentation

Show task log

Description

Writes the task log for the given task

Usage

taskLog(taskId, account = NULL, server = NULL, output = NULL)

Arguments

taskId
Task Id

account, server
Uniquely identify a remote server with either your user account, the server name, or both. If neither are supplied, and there are multiple options, you'll be prompted to pick one.

Use accounts() to see the full list of available options.

output
Where to write output. Valid values are NULL or stderr

See Also

tasks()

Examples

## Not run: 

# write task log to stdout
taskLog(12345)

# write task log to stderr
taskLog(12345, output="stderr")

## End(Not run)

[Package rsconnect version 1.2.1 Index]