tasks {rsconnect}	R Documentation

List Tasks

Description

List Tasks

Usage

tasks(account = NULL, server = NULL)

Arguments

account, server
Uniquely identify a remote server with either your user account, the server name, or both. If neither are supplied, and there are multiple options, you'll be prompted to pick one.

Use accounts() to see the full list of available options.

Value

Returns a data frame with the following columns:

id 	 Task id

action 	 Task action

status 	 Current task status

created_time 	 Task creation time

finished_time 	 Task finished time

See Also

taskLog()

Examples

## Not run: 

# list tasks for the default account
tasks()

## End(Not run)

[Package rsconnect version 1.2.1 Index]