terminateApp {rsconnect}	R Documentation

Terminate an Application

Description

Terminate and archive a currently deployed ShinyApps application.

Usage

terminateApp(appName, account = NULL, server = NULL, quiet = FALSE)

Arguments

appName
Name of application to terminate

account
Account name. If a single account is registered on the system then this parameter can be omitted.

server
Server name. Required only if you use the same account name on multiple servers (see servers())

quiet
Request that no status information be printed to the console during the termination.

Note

This function only works for ShinyApps servers.

See Also

applications(), deployApp(), and restartApp()

Examples

## Not run: 

# terminate an application
terminateApp("myapp")

## End(Not run)

[Package rsconnect version 0.8.26 Index]