unsetProperty {rsconnect}	R Documentation

Unset Application property

Description

Unset a property on currently deployed ShinyApps application (restoring to its default value)

Usage

unsetProperty(
  propertyName,
  appPath = getwd(),
  appName = NULL,
  account = NULL,
  force = FALSE
)

Arguments

propertyName
Name of property to unset

appPath
Directory or file that was deployed. Defaults to current working directory.

appName
Name of application

account
Account name. If a single account is registered on the system then this parameter can be omitted.

force
Forcibly unset the property

Note

This function only works for ShinyApps servers.

Examples

## Not run: 

# unset application package cache property to revert to default
unsetProperty("application.package.cache")

## End(Not run)

[Package rsconnect version 0.8.26 Index]