addAuthorizedUser {rsconnect}	R Documentation

Add authorized user to application

Description

Add authorized user to application

Usage

addAuthorizedUser(
  email,
  appDir = getwd(),
  appName = NULL,
  account = NULL,
  server = NULL,
  sendEmail = NULL,
  emailMessage = NULL
)

Arguments

email
Email address of user to add.

appDir
Directory containing application. Defaults to current working directory.

appName
Name of application.

account
Account name. If a single account is registered on the system then this parameter can be omitted.

server
Server name. Required only if you use the same account name on multiple servers.

sendEmail
Send an email letting the user know the application has been shared with them.

emailMessage
Optional character vector of length 1 containing a custom message to send in email invitation. Defaults to NULL, which will use default invitation message.

Note

This function works only for ShinyApps servers.

See Also

removeAuthorizedUser() and showUsers()

[Package rsconnect version 0.8.26 Index]