configureApp {rsconnect}	R Documentation

Configure an Application

Description

Configure an application running on a remote server.

Usage

configureApp(
  appName,
  appDir = getwd(),
  account = NULL,
  server = NULL,
  redeploy = TRUE,
  size = NULL,
  instances = NULL,
  logLevel = c("normal", "quiet", "verbose")
)

Arguments

appName
Name of application to configure

appDir
Directory containing application. Defaults to current working directory.

account
Account name. If a single account is registered on the system then this parameter can be omitted.

server
Server name. Required only if you use the same account name on multiple servers (see servers())

redeploy
Re-deploy application after its been configured.

size
Configure application instance size

instances
Configure number of application instances

logLevel
One of "quiet", "normal" or "verbose"; indicates how much logging to the console is to be performed. At "quiet" reports no information; at "verbose", a full diagnostic log is captured.

Note

This function works only for ShinyApps servers.

See Also

applications(), deployApp()

Examples

## Not run: 

# set instance size for an application
configureApp("myapp", size="xlarge")

## End(Not run)

[Package rsconnect version 0.8.26 Index]