connectApiUser {rsconnect}	R Documentation

Connect Api User Account

Description

Connect a user account to the package using an API key for authentication so that it can be used to deploy and manage applications on behalf of the account.

Usage

connectApiUser(account = NULL, server = NULL, apiKey = NULL, quiet = FALSE)

Arguments

account
A name for the account to connect. Optional.

server
The server to connect to. Optional if there is only one server registered.

apiKey
The API key used to authenticate the user

quiet
Whether or not to show messages and prompts while connecting the account.

Details

This function configures the user to connect using an apiKey in the http auth headers instead of a token. This is less secure but may be necessary when the client is behind a proxy or otherwise unable to authenticate using a token.

See Also

Other Account functions: accounts(), connectUser(), setAccountInfo()

[Package rsconnect version 0.8.26 Index]