deployAPI {rsconnect}	R Documentation

Deploy a Plumber API

Description

Deploys an application consisting of plumber API routes. The given directory must contain a script returning a plumb object or a plumber API definition.

Usage

deployAPI(api, ...)

Arguments

api
Path to the API project directory. Must contain either entrypoint.R or plumber.R

...
Additional arguments to deployApp().

Details

Deploy a plumber API definition by either supplying a directory containing plumber.R (an API definition) or entrypoint.R that returns a plumb object created by plumber::plumb(). See the plumber documentation for more information.

See Also

Other Deployment functions: applications(), deployApp(), deployDoc(), deploySite(), deployTFModel()

[Package rsconnect version 0.8.26 Index]