deployDoc {rsconnect}	R Documentation

Deploy a Document

Description

Deploys an application consisting of a single R Markdown document or other single file (such as an HTML or PDF document).

Usage

deployDoc(doc, ...)

Arguments

doc
Path to the document to deploy.

...
Additional arguments to deployApp(). Do not supply appDir, appFiles, or appPrimaryDoc; these three parameters are automatically generated by deployDoc from the document.

Details

When deploying an R Markdown document, any files which are required to render and display the file must be deployed.

This method discovers these additional files using rmarkdown::find_external_resources() from rmarkdown.

If you find that the document is missing dependencies, either specify the dependencies explicitly in the document (the documentation for rmarkdown::find_external_resources() explains how to do this), or call deployApp() directly and specify your own file list in the appFiles parameter.

See Also

Other Deployment functions: applications(), deployAPI(), deployApp(), deploySite(), deployTFModel()

[Package rsconnect version 0.8.26 Index]