deploySite {rsconnect}	R Documentation

Deploy a Website

Description

Deploy an R Markdown website to a server.

Usage

deploySite(
  siteDir = getwd(),
  siteName = NULL,
  account = NULL,
  server = NULL,
  render = c("none", "local", "server"),
  launch.browser = getOption("rsconnect.launch.browser", interactive()),
  logLevel = c("normal", "quiet", "verbose"),
  lint = FALSE,
  metadata = list(),
  python = NULL,
  ...
)

Arguments

siteDir
Directory containing website. Defaults to current working directory.

siteName
Name for the site (names must be unique within an account). Defaults to the base name of the specified siteDir, (or to a name provided by a custom site generation function).

account
Account to deploy application to. This parameter is only required for the initial deployment of an application when there are multiple accounts configured on the system (see accounts).

server
Server name. Required only if you use the same account name on multiple servers.

render
Rendering behavior for site: "none" to upload a static version of the current contents of the site directory; "local" to render the site locally then upload it; "server" to render the site on the server. Note that for "none" and "local" R scripts (.R) and markdown documents (.Rmd and .md) will not be uploaded to the server.

launch.browser
If true, the system's default web browser will be launched automatically after the app is started. Defaults to TRUE in interactive sessions only. If a function is passed, it will be called after the app is started, with the app URL as a paramter.

logLevel
One of "quiet", "normal" or "verbose"; indicates how much logging to the console is to be performed. At "quiet" reports no information; at "verbose", a full diagnostic log is captured.

lint
Lint the project before initiating deployment, to identify potentially problematic code?

metadata
Additional metadata fields to save with the deployment record. These fields will be returned on subsequent calls to deployments().

python
Full path to a python binary for use by reticulate. Required if reticulate is a dependency of the app being deployed. If python = NULL, and RETICULATE_PYTHON is set in the environment, its value will be used. The specified python binary will be invoked to determine its version and to list the python packages installed in the environment.

...
Additional arguments to deployApp(). Do not supply appDir, appFiles, or appSourceDoc; these three parameters are automatically generated by deploySite.

See Also

Other Deployment functions: applications(), deployAPI(), deployApp(), deployDoc(), deployTFModel()

[Package rsconnect version 0.8.26 Index]