List Application Deployments

List deployment records for a given application.

deployments(
  appPath,
  nameFilter = NULL,
  accountFilter = NULL,
  serverFilter = NULL,
  excludeOrphaned = TRUE
)

Arguments

appPath
The path to the content that was deployed, either a directory or an individual document.

nameFilter
Return only deployments matching the given name (optional)

accountFilter
Return only deployments matching the given account (optional)

serverFilter
Return only deployments matching the given server (optional)

excludeOrphaned
If TRUE (the default), return only deployments made by a currently registered account. Deployments made from accounts that are no longer registered (via e.g.removeAccount()) will not be returned.

Value

Returns a data frame with at least following columns:

name	Name of deployed application

account	Account owning deployed application

bundleId	Identifier of deployed application's bundle

url	URL of deployed application

when	When the application was deployed (in seconds since the epoch)

If additional metadata has been saved with the deployment record using the metadata argument to deployApp(), the frame will include additional columns.

See also

Examples

if (FALSE) {

# Return all deployments of the ~/r/myapp directory made with the 'abc'
# account
deployments("~/r/myapp", accountFilter="abc")
}