listBundleFiles {rsconnect}	R Documentation

List Files to be Bundled

Description

Given a directory containing an application, returns the names of the files to be bundled in the application.

Usage

listBundleFiles(appDir)

Arguments

appDir
Directory containing the application.

Details

This function computes results similar to a recursive directory listing from list.files(), with the following constraints:

If the total size of the files exceeds the maximum bundle size, no more files are listed. The maximum bundle size is controlled by the rsconnect.max.bundle.size option.

If the total size number of files exceeds the maximum number to be bundled, no more files are listed. The maximum number of files in the bundle is controlled by the rsconnect.max.bundle.files option.

Certain files and folders that don't need to be bundled, such as those containing internal version control and RStudio state, are excluded.

In order to stop specific files in the working directory from being listed in the bundle, the files must be listed in the .rscignore file. This file must have one file or directory per line with no support for wildcards.

Value

Returns a list containing the following elements:

contents 	 A list of the files to be bundled

totalSize 	 The total size of the files

[Package rsconnect version 0.8.26 Index]