removeAuthorizedUser {rsconnect}	R Documentation

Remove authorized user from an application

Description

Remove authorized user from an application

Usage

removeAuthorizedUser(
  user,
  appDir = getwd(),
  appName = NULL,
  account = NULL,
  server = NULL
)

Arguments

user
The user to remove. Can be id or email address.

appDir
Directory containing application. Defaults to current working directory.

appName
Name of application.

account
Account name. If a single account is registered on the system then this parameter can be omitted.

server
Server name. Required only if you use the same account name on multiple servers.

Note

This function works only for ShinyApps servers.

See Also

addAuthorizedUser() and showUsers()

[Package rsconnect version 0.8.26 Index]