servers {rsconnect}	R Documentation

Server metadata

Description

servers() lists all known servers; serverInfo() gets metadata about a specific server. Cloud servers shinyapps.io and posit.cloud are always automatically registered and available.

Usage

servers(local = FALSE)

serverInfo(name = NULL)

Arguments

local
Return only local servers? (i.e. not automatically registered cloud servers)

name
Server name. If omitted, you'll be prompted to pick a server.

Value

servers() returns a data frame with registered server names and URLs. serverInfo() returns a list with details for a particular server.

Examples

# List all registered servers
servers()

# Get information about a server
serverInfo("posit.cloud")

[Package rsconnect version 1.2.1 Index]