taskLog {rsconnect}	R Documentation

Show task log

Description

Writes the task log for the given task

Usage

taskLog(taskId, account = NULL, server = NULL, output = NULL)

Arguments

taskId
Task Id

account
Account name. If a single account is registered on the system then this parameter can be omitted.

server
Server name. Required only if you use the same account name on multiple servers (see servers())

output
Where to write output. Valid values are NULL or stderr

See Also

tasks()

Examples

## Not run: 

# write task log to stdout
taskLog(12345)

# write task log to stderr
taskLog(12345, output="stderr")

## End(Not run)

[Package rsconnect version 0.8.26 Index]