writeManifest {rsconnect}	R Documentation

Create a manifest.json describing deployment requirements.

Description

Given a directory content targeted for deployment, write a manifest.json into that directory describing the deployment requirements for that content.

Usage

writeManifest(appDir = getwd(), appFiles = NULL,
  appPrimaryDoc = NULL, contentCategory = NULL)

Arguments

appDir
Directory containing the content (Shiny application, R Markdown document, etc).

appFiles
Optional. The full set of files and directories to be included in future deployments of this content. Used when computing dependency requirements. When NULL, all files in appDir are considered.

appPrimaryDoc
Optional. Specifies the primary document in a content directory containing more than one. If NULL, the primary document is inferred from the file list.

contentCategory
Optional. Specifies the kind of content being deployed (e.g. "plot" or "site").

[Package rsconnect version 0.8.11 Index]