writeManifest {rsconnect}	R Documentation

Create a manifest.json describing deployment requirements.

Description

Given a directory content targeted for deployment, write a manifest.json into that directory describing the deployment requirements for that content.

Usage

writeManifest(
  appDir = getwd(),
  appFiles = NULL,
  appPrimaryDoc = NULL,
  contentCategory = NULL,
  python = NULL,
  forceGeneratePythonEnvironment = FALSE,
  verbose = FALSE
)

Arguments

appDir
Directory containing the content (Shiny application, R Markdown document, etc).

appFiles
Optional. The full set of files and directories to be included in future deployments of this content. Used when computing dependency requirements. When NULL, all files in appDir are considered.

appPrimaryDoc
Optional. Specifies the primary document in a content directory containing more than one. If NULL, the primary document is inferred from the file list.

contentCategory
Optional. Specifies the kind of content being deployed (e.g. "plot" or "site").

python
Full path to a python binary for use by reticulate. The specified python binary will be invoked to determine its version and to list the python packages installed in the environment. If python = NULL, and RETICULATE_PYTHON is set in the environment, its value will be used.

forceGeneratePythonEnvironment
Optional. If an existing requirements.txt file is found, it will be overwritten when this argument is TRUE.

verbose
If TRUE, prints progress messages to the console

[Package rsconnect version 0.8.26 Index]